feat: add abstracting the public React component from init
