+++ apps/mis-web/src/pages/init/importUsers.tsx
@@ -1,31 +1,17 @@
-import { Button, Modal, Result, Tabs } from "antd";
+import { Result, Tabs } from "antd";
 import { GetServerSideProps, NextPage } from "next";
-import Router from "next/router";
-import { api } from "src/apis";
+import Link from "next/link";
 import { SSRProps } from "src/auth/server";
 import { UnifiedErrorPage } from "src/components/errorPages/UnifiedErrorPage";
-import { Centered } from "src/components/layouts";
 import { EditJobPriceTableForm } from "src/pageComponents/init/EditJobPriceTableForm";
 import { InitAdminForm } from "src/pageComponents/init/InitAdminForm";
 import { InitImportUsersForm } from "src/pageComponents/init/InitImportUsersForm";
-import { InitUsersAndAccountsTable } from "src/pageComponents/init/InitUsersAndAccountsTable";
-import { Head } from "src/utils/head";
+import { InitDrawer } from "src/pageComponents/init/InitLayout";
+// import { InitUsersAndAccountsTable } from "src/pageComponents/init/InitUsersAndAccountsTable";
 import { queryIfInitialized } from "src/utils/init";
-import styled from "styled-components";
-
 type Props = SSRProps<{}>;
 
-const Title = styled(Centered)`
-  position: relative;
-`;
-
-const CompleteButtonContainer = styled.div`
-  position: absolute;
-  right: 0;
-`;
-
-export const InitSystemPage: NextPage<Props> = (props) => {
-
+export const ImportUsersPage: NextPage<Props> = (props) => {
   if ("error" in props) {
     return (
       <UnifiedErrorPage code={props.error}
@@ -41,40 +27,15 @@ export const InitSystemPage: NextPage<Props> = (props) => {
       />
     );
   }
-
   return (
     <div>
-      <Head title="系统初始化"/>
-      <Title>
-        <span />
-        <h1>系统初始化</h1>
-        <CompleteButtonContainer>
-          <Button type="primary" onClick={() => {
-            Modal.confirm({
-              title: "确认完成初始化",
-              content: "一旦完成初始化，您将无法进入此页面重新初始化。",
-              onOk: () => api.completeInit({}).then(() => {
-                Modal.success({
-                  title: "初始化完成！",
-                  content: "点击确认前往登录",
-                  closable: false,
-                  maskClosable: false,
-                  onOk: () => Router.push("/api/auth"),
-                });
-              }),
-            });
-          }}
-          >
-          完成初始化
-          </Button>
-        </CompleteButtonContainer>
-      </Title>
+      <InitDrawer/>
       <Tabs centered defaultActiveKey="1">
         <Tabs.TabPane tab="导入用户" key="1">
           <InitImportUsersForm />
         </Tabs.TabPane>
         <Tabs.TabPane tab="用户账户管理" key="2">
-          <InitUsersAndAccountsTable />
+          <Link href="/users"/>
         </Tabs.TabPane>
         <Tabs.TabPane tab="创建初始管理员用户" key="3">
           <InitAdminForm />
@@ -98,4 +59,4 @@ export const getServerSideProps: GetServerSideProps<Props> = async () => {
 
 };
 
-export default InitSystemPage;
+export default ImportUsersPage;
