+++ apps/mis-web/src/pages/init/users.tsx
@@ -1,0 +1,1 @@
+console.log("i'm flag");
